# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Returns Services API",
        "description": "API documentation for Returns Service App",
        "contact": {
            "email": "admin@adm.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://localhost/api/",
            "description": "Returns App API"
        }
    ],
    "paths": {
...
        "/pemeriksaans": {
            "get": {
                "tags": [
                    "Pemeriksaan"
                ],
                "summary": "getPemeriksaanList",
                "description": "Get all Pemeriksaans",
                "operationId": "561a9cd31418a0cb78b6feddc8717d9d",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Pemeriksaan"
                ],
                "summary": "createPemeriksaan",
                "description": "Create Pemeriksaan",
                "operationId": "d0c10f16dfb3699fd4a840a2e1e9e241",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "desc",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/pemeriksaans/{id}": {
            "get": {
                "tags": [
                    "Pemeriksaan"
                ],
                "summary": "getPemeriksaanItem",
                "description": "Get Pemeriksaan",
                "operationId": "25549947a4b73e9de93ae518e557b6c9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Pemeriksaan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pemeriksaan"
                ],
                "summary": "updatePemeriksaan",
                "description": "Update Pemeriksaan",
                "operationId": "0dd28a9bd2f0275cd5882042f3d1fa6b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Pemeriksaan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "desc",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pemeriksaan"
                ],
                "summary": "deletePemeriksaan",
                "description": "Delete Pemeriksaan",
                "operationId": "9449c7eaa7117aa091a382f11cc69189",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Pemeriksaan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
...
}